feat(kuzu): support kuzu as storage target (#561)
* feat(kuzu): support setup

* style: rearrange `kuzu.rs` into sections

* refactor(kuzu): expose single `register()` method to seal internal types

* refactor: use standalone methods to build setup query

* feat(kuzu): support kuzu as target

+++ src/base/schema.rs
@@ -86,7 +86,7 @@ impl std::fmt::Display for BasicValueType {
     }
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Default)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default)]
 pub struct StructSchema {
     pub fields: Arc<Vec<FieldSchema>>,
 
@@ -138,7 +138,7 @@ impl std::fmt::Display for TableKind {
     }
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 pub struct TableSchema {
     pub kind: TableKind,
     pub row: StructSchema,
@@ -191,7 +191,7 @@ impl TableSchema {
     }
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 #[serde(tag = "kind")]
 pub enum ValueType {
     Struct(StructSchema),
@@ -222,7 +222,7 @@ impl ValueType {
     }
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 pub struct EnrichedValueType<DataType = ValueType> {
     #[serde(rename = "type")]
     pub typ: DataType,
@@ -295,7 +295,7 @@ impl std::fmt::Display for ValueType {
     }
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 pub struct FieldSchema<DataType = ValueType> {
     /// ID is used to identify the field in the schema.
     pub name: FieldName,
@@ -342,7 +342,7 @@ impl std::fmt::Display for FieldSchema {
 pub struct CollectorSchema {
     pub fields: Vec<FieldSchema>,
     /// If specified, the collector will have an automatically generated UUID field with the given index.
-    pub auto_uuid_field_idx: Option<u32>,
+    pub auto_uuid_field_idx: Option<usize>,
 }
 
 impl std::fmt::Display for CollectorSchema {
